[Scale] Automatically close Scale if opened on an empty workspace.
It makes little sense to open Scale on an empty workspace, and it can
be bewildering even to experienced users to see nothing
but a dimmed background image, especially if they opened
Scale by mistake.

This patch makes Scale briefly display a message and return
to the current desktop, if opened on an empty workspace.